EES-4817 Move DataSetGeneratorExtensions to `Public.Data.Model.Tests`
